requirements.txt
click==8.1.7
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==1.26.4
opencv-python==4.10.0.82
pillow==10.3.0
pygments==2.18.0
rich==13.7.1
shellingham==1.5.4
typer==0.12.3
typing-extensions==4.12.1
